requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
chardet==4.0.0
discord.py==1.7.3
idna==3.2
multidict==5.1.0
pycodestyle==2.7.0
python-dotenv==0.19.0
toml==0.10.2
typing-extensions==3.10.0.0
yarl==1.6.3